Refactor current downshift integration demo
Extract out common constants and functions
